+++ gradlew.bat
@@ -10,7 +10,7 @@
 @rem Unless required by applicable law or agreed to in writing, software
 @rem distributed under the License is distributed on an "AS IS" BASIS,
 @rem WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-@rem See the License for the specific language governing permissions and
+@rem See the License for the specific languageType governing permissions and
 @rem limitations under the License.
 @rem
 @rem SPDX-License-Identifier: Apache-2.0
