+++ src/main/java/org/example/sqi_images/employee/dto/response/ImageDataResponse.java
@@ -1,4 +1,4 @@
-package org.example.sqi_images.profile.dto.response;
+package org.example.sqi_images.employee.dto.response;
 
 public record ImageDataResponse(byte[] imageData, String contentType) {
     public static ImageDataResponse of(byte[] imageData, String contentType) {
