+++ src/main/java/org/example/sqi_images/employee/dto/response/ProfileResponse.java
@@ -1,4 +1,4 @@
-package org.example.sqi_images.profile.dto.response;
+package org.example.sqi_images.employee.dto.response;
 
 import jakarta.persistence.Lob;
 import lombok.Builder;
